moved legacy examples
