+++ src/legacy/examples/custom-instance.js
@@ -1,5 +1,5 @@
 //@ts-check
-const { CustomInstance } = require('../src/better-logging');
+const { CustomInstance } = require('../../better-logging');
 
 const logs = [];
 const removeColor = msg => msg.replace(/[\u001b\u009b][[()#;?]*(?:[0-9]{1,4}(?:;[0-9]{0,4})*)?[0-9A-ORZcf-nqry=><]/g, '');
