+++ src/legacy/examples/custom-stamp-color.js
@@ -1,6 +1,6 @@
 //@ts-check
 const better = {};
-if (!require('../src/better-logging').default(better, {
+if (!require('../../better-logging').default(better, {
     stampColor: Color => Color.Light_Green
 })) throw 'This will never happen';
