+++ src/legacy/examples/decorate-object.js
@@ -1,6 +1,6 @@
 //@ts-check
 const better = {};
-if (!require('../src/better-logging').default(better)) throw 'This will never happen';
+if (!require('../../better-logging').default(better)) throw 'This will never happen';
 
 Array(5).fill(0).forEach((_, i) => {
   better.loglevel = i;
